# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasyStay Documentation",
        "description": "Ceci est la documentation de EasyStay.",
        "contact": {
            "email": "haroldavademe0@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion Utilisateur",
                "description": "Connexion d'un utilisateur et retour d'un token",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Adresse email de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "user@example.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Mot de passe de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password",
                            "example": "password123"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Identifiants de connexion (facultatif si fournis en paramètres)",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Adresse email de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Indique s'il y a une erreur",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Message d'authentification",
                                            "type": "string",
                                            "example": "Vous êtes authentifié"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Sudo"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "sudo@admin.hotel"
                                                },
                                                "adresse": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "tel": {
                                                    "type": "string",
                                                    "example": "+229 91461545"
                                                },
                                                "picture": {
                                                    "type": "string",
                                                    "example": "images/blank_profile.jpeg"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "Sudo"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Enable"
                                                },
                                                "hotels_id": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-25T01:08:30.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-25T01:08:30.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "5|2YaDKuz5UPdqAR2RoVOb9sxbz92a9b5JJGvGzgGs29e54f58"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Identifiants invalides"
                    },
                    "422": {
                        "description": "Erreur de validation - Champs requis manquants"
                    }
                }
            }
        },
...
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout the authenticated user",
                "description": "Deletes the current access token and logs out the user.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Indicates if there was an error",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "Déconnexion réussite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}